Move the tsc helpers in one place
