+++ src/testRunner/unittests/tsbuild/exitCodeOnBogusFile.ts
@@ -1,5 +1,4 @@
-import { verifyTsc } from "../tsc/helpers";
-import { loadProjectFromFiles } from "./helpers";
+import { loadProjectFromFiles, verifyTsc } from "../tsc/helpers";
 
 // https://github.com/microsoft/TypeScript/issues/33849
 describe("unittests:: tsbuild:: exitCodeOnBogusFile:: test exit code", () => {
