+++ src/testRunner/unittests/tsbuild/moduleSpecifiers.ts
@@ -1,7 +1,6 @@
 import { libFile } from "../../../harness/virtualFileSystemWithWatch";
 import * as Utils from "../../_namespaces/Utils";
-import { verifyTsc } from "../tsc/helpers";
-import { loadProjectFromFiles, symbolLibContent } from "./helpers";
+import { loadProjectFromFiles, symbolLibContent, verifyTsc } from "../tsc/helpers";
 
 // https://github.com/microsoft/TypeScript/issues/31696
 describe("unittests:: tsbuild:: moduleSpecifiers:: synthesized module specifiers to referenced projects resolve correctly", () => {
