+++ src/testRunner/unittests/tsbuild/noEmit.ts
@@ -1,5 +1,4 @@
-import { noChangeRun } from "../tsc/helpers";
-import { loadProjectFromFiles, verifyTscWithEdits } from "./helpers";
+import { loadProjectFromFiles, noChangeRun, verifyTscWithEdits } from "../tsc/helpers";
 
 describe("unittests:: tsbuild:: noEmit", () => {
     function verifyNoEmitWorker(subScenario: string, aTsContent: string, commandLineArgs: readonly string[]) {
