+++ src/testRunner/unittests/tsbuild/referencesWithRootDirInParent.ts
@@ -1,6 +1,5 @@
 import * as vfs from "../../_namespaces/vfs";
-import { verifyTsc } from "../tsc/helpers";
-import { loadProjectFromDisk, replaceText } from "./helpers";
+import { loadProjectFromDisk, replaceText, verifyTsc } from "../tsc/helpers";
 
 describe("unittests:: tsbuild:: with rootDir of project reference in parentDirectory", () => {
     let projFs: vfs.FileSystem;
