+++ src/testRunner/unittests/tsbuild/transitiveReferences.ts
@@ -1,6 +1,5 @@
 import * as vfs from "../../_namespaces/vfs";
-import { verifyTsc } from "../tsc/helpers";
-import { loadProjectFromDisk } from "./helpers";
+import { loadProjectFromDisk, verifyTsc } from "../tsc/helpers";
 
 describe("unittests:: tsbuild:: when project reference is referenced transitively", () => {
     let projFs: vfs.FileSystem;
