+++ src/testRunner/unittests/tsbuildWatch/noEmit.ts
@@ -1,6 +1,6 @@
 import { createWatchedSystem, libFile } from "../../../harness/virtualFileSystemWithWatch";
+import { libContent } from "../tsc/helpers";
 import { verifyTscWatch } from "../tscWatch/helpers";
-import { libContent } from "../tsbuild/helpers";
 
 describe("unittests:: tsbuildWatch:: watchMode:: with noEmit", () => {
     verifyTscWatch({
