+++ src/testRunner/unittests/tsc/forceConsistentCasingInFileNames.ts
@@ -1,6 +1,5 @@
 import * as Utils from "../../_namespaces/Utils";
-import { loadProjectFromFiles } from "../tsbuild/helpers";
-import { verifyTsc } from "./helpers";
+import { loadProjectFromFiles, verifyTsc } from "./helpers";
 
 describe("unittests:: tsc:: forceConsistentCasingInFileNames::", () => {
     verifyTsc({
