+++ src/testRunner/unittests/tsc/listFilesOnly.ts
@@ -1,6 +1,5 @@
 import * as Utils from "../../_namespaces/Utils";
-import { loadProjectFromFiles, verifyTscWithEdits } from "../tsbuild/helpers";
-import { noChangeRun, verifyTsc } from "./helpers";
+import { loadProjectFromFiles, noChangeRun, verifyTsc, verifyTscWithEdits } from "./helpers";
 
 describe("unittests:: tsc:: listFilesOnly::", () => {
     verifyTsc({
